feat: 🎸 add new fields to property fn
