+++ lua/luasnip-ts-snippets/luasnippets/python/snip_functions.lua
@@ -59,29 +59,6 @@ local function snip_node(desc, decorator, ref)
    )
 end
 
-local function property_snip_node(desc, lookup_key)
-   return l.sn(
-      nil,
-      l.fmta(py_utils.property_declarations[lookup_key], {
-         name = l.i(1, "foo"),
-         retval = l.i(2, "None"),
-         getter_body = l.d(
-            3,
-            function(args) return l.sn(nil, l.i(1, string.format("return self._%s", args[1][1]))) end,
-            { 1 }
-         ),
-         rep = l.rep(1),
-         value = l.i(4, "value"),
-         setter_body = l.d(
-            5,
-            function(args) return l.sn(nil, l.i(1, string.format("self._%s = %s", args[1][1], args[2][1]))) end,
-            { 1, 2 }
-         ),
-      }, { strict = false }),
-      u.desc(desc)
-   )
-end
-
 local main_fn = [[
 if __name__ == "__main__":
     <body>
@@ -100,18 +77,6 @@ return {
          snip_node("Method", nil, "self"),
          snip_node("Staticmethod", "@staticmethod", nil),
          snip_node("Classmethod", "@classmethod", "cls"),
-      })
-   ),
-
-   l.s(
-      {
-         trig = "property",
-         name = "Property boilerplate",
-         dscr = "Create new property",
-      },
-      l.c(1, {
-         property_snip_node("Property getter", "getter"),
-         property_snip_node("Property getter+setter", "getter_setter"),
       })
    ),
